feat(http): remove HTTP client implementation and related modules
- Removed entire http_client.rs module with HttpClient and HttpConfig implementations
- Deleted all HTTP request functionality including fetch, get, post, put, delete methods
- Removed retry mechanism and timeout control features
- Eliminated proxy configuration and rate limiting integration
- Removed circuit breaker support and gzip compression/decompression
- Deleted request/response logging and structured tracing functionality
- Removed all associated tests for HTTP client functionality
- Removed dependencies on reqwest, serde_json and other HTTP-related crates
